code clean up.
