+++ timbuctoo-instancev4/src/main/scripts/index_scripts/dcar/configs/dcar_archive_config.rb
@@ -9,8 +9,8 @@ class DcarArchiveConfig
         { :name => 'titleNld',  :converted_name => 'titleNld_t'},
         { :name => 'documentType', :converted_name => 'documentType_s'},
         { :name => 'notes', :converted_name => 'notes_t'},
-        { :name => 'beginDate', :converted_name => 'beginDate_dt', :type => 'date'},
-        { :name => 'endDate', :converted_name => 'endDate_dt', :type => 'date'},
+        { :name => 'beginDate', :converted_name => 'beginDate_dt', :type => 'dcar_date'},
+        { :name => 'endDate', :converted_name => 'endDate_dt', :type => 'dcar_date'},
         { :name => 'beginDate', :converted_name => 'beginDate_i', :type => 'year'},
         { :name => 'endDate', :converted_name => 'endDate_i', :type => 'year'},
         { :name => 'refCodeArchive', :converted_name => 'refCodeArchive_s'},
